+++ app/src/main/java/com/oreocube/booksearch/MainNavHost.kt
@@ -10,6 +10,7 @@ import com.oreocube.booksearch.feature.book.detail.bookDetailScreen
 import com.oreocube.booksearch.feature.book.detail.navigateToBookDetail
 import com.oreocube.booksearch.feature.book.search.barcode.BookBarcodeScannerRoute
 import com.oreocube.booksearch.feature.book.search.barcode.bookBarcodeScannerScreen
+import com.oreocube.booksearch.feature.book.search.barcode.navigateToBookBarcodeScanner
 import com.oreocube.booksearch.feature.book.search.navigateToSearchBook
 import com.oreocube.booksearch.feature.book.search.searchBookScreen
 import com.oreocube.booksearch.feature.discovery.discoveryScreen
@@ -54,6 +55,7 @@ fun MainNavHost(
         searchBookScreen(
             onBackClick = navController::popBackStack,
             onBookClick = navController::navigateToBookDetail,
+            onBarcodeScanClick = navController::navigateToBookBarcodeScanner,
             onShowSnackbar = onShowSnackbar,
         )
         bookDetailScreen(
