+++ core/ui/src/main/res/drawable/ic_barcode_scanner_24.xml
@@ -1,0 +1,5 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:tint="#000000" android:viewportHeight="960" android:viewportWidth="960" android:width="24dp">
+      
+    <path android:fillColor="@android:color/white" android:pathData="M240,800Q240,817 228.5,828.5Q217,840 200,840L80,840Q63,840 51.5,828.5Q40,817 40,800L40,680Q40,663 51.5,651.5Q63,640 80,640Q97,640 108.5,651.5Q120,663 120,680L120,760L200,760Q217,760 228.5,771.5Q240,783 240,800ZM880,640Q897,640 908.5,651.5Q920,663 920,680L920,800Q920,817 908.5,828.5Q897,840 880,840L760,840Q743,840 731.5,828.5Q720,817 720,800Q720,783 731.5,771.5Q743,760 760,760L840,760L840,680Q840,663 851.5,651.5Q863,640 880,640ZM180,720Q172,720 166,714Q160,708 160,700L160,260Q160,252 166,246Q172,240 180,240L220,240Q228,240 234,246Q240,252 240,260L240,700Q240,708 234,714Q228,720 220,720L180,720ZM300,720Q292,720 286,714Q280,708 280,700L280,260Q280,252 286,246Q292,240 300,240Q308,240 314,246Q320,252 320,260L320,700Q320,708 314,714Q308,720 300,720ZM420,720Q412,720 406,714Q400,708 400,700L400,260Q400,252 406,246Q412,240 420,240L460,240Q468,240 474,246Q480,252 480,260L480,700Q480,708 474,714Q468,720 460,720L420,720ZM540,720Q532,720 526,714Q520,708 520,700L520,260Q520,252 526,246Q532,240 540,240L620,240Q628,240 634,246Q640,252 640,260L640,700Q640,708 634,714Q628,720 620,720L540,720ZM700,720Q692,720 686,714Q680,708 680,700L680,260Q680,252 686,246Q692,240 700,240Q708,240 714,246Q720,252 720,260L720,700Q720,708 714,714Q708,720 700,720ZM780,720Q772,720 766,714Q760,708 760,700L760,260Q760,252 766,246Q772,240 780,240Q788,240 794,246Q800,252 800,260L800,700Q800,708 794,714Q788,720 780,720ZM240,160Q240,177 228.5,188.5Q217,200 200,200L120,200L120,280Q120,297 108.5,308.5Q97,320 80,320Q63,320 51.5,308.5Q40,297 40,280L40,160Q40,143 51.5,131.5Q63,120 80,120L200,120Q217,120 228.5,131.5Q240,143 240,160ZM720,160Q720,143 731.5,131.5Q743,120 760,120L880,120Q897,120 908.5,131.5Q920,143 920,160L920,280Q920,297 908.5,308.5Q897,320 880,320Q863,320 851.5,308.5Q840,297 840,280L840,200L760,200Q743,200 731.5,188.5Q720,177 720,160Z"/>
+    
+</vector>
